+++ sample/src/commonMain/kotlin/com/kez/picker/sample/ui/screen/IntegratedPickerScreen.kt
@@ -153,6 +153,7 @@ internal fun IntegratedPickerScreen(
                             if (tabIndex == 0) {
                                 YearMonthPicker(
                                     state = yearMonthState,
+                                    spacingBetweenPickers = 8.dp,
                                     format = PickerDefaults.yearMonthPickerFormat(
                                         yearItemText = { "${it}년" },
                                         monthItemText = { getMonthName(it) },
@@ -185,6 +186,7 @@ internal fun IntegratedPickerScreen(
                             } else {
                                 TimePicker(
                                     state = timeState,
+                                    spacingBetweenPickers = 8.dp,
                                     format = PickerDefaults.timePickerFormat(
                                         hourItemText = { it.toString().padStart(2, '0') },
                                         minuteItemText = { it.toString().padStart(2, '0') },